requirements.txt
streamlit>=1.31.0
anthropic>=0.18.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
